feat(#8): Add vault/unvault action
